requirements.txt
beautifulsoup4
requests
streamlit
python-safer